ADC sampling rate as long.

+++ src/drivers/drivers/adc.h
@@ -56,7 +56,7 @@ int adc_init(struct adc_driver_t *self_p,
              struct adc_device_t *dev_p,
              struct pin_device_t *pin_dev_p,
              int reference,
-             int sampling_rate);
+             long sampling_rate);
 
 /**
  * Start an asynchronous convertion of analog signal to digital
